# server/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ShareFridge API",
    "version": "2.0.0",
    "description": "Room-authenticated remediation contract. x-availability marks future integrations; these must not be claimed ready."
  },
  "paths": {
...
    "/api/openapi.json": {
      "get": {
        "summary": "GET /api/openapi.json",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "openapi",
                    "info",
                    "paths",
                    "components"
                  ],
                  "properties": {
                    "openapi": {
                      "type": "string"
                    },
                    "info": {
                      "type": "object"
                    },
                    "paths": {
                      "type": "object"
                    },
                    "components": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Invalid session/credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Foreign room",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Missing item",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Oversized request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/config": {
      "get": {
        "summary": "GET /api/config",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicConfig"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Invalid session/credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Foreign room",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Missing item",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Oversized request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/create-room": {
      "post": {
        "summary": "POST /api/auth/create-room",
        "security": [],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSession"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Invalid session/credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Foreign room",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Missing item",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Oversized request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomCreate"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error",
          "code"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      },
...
      "PublicConfig": {
        "type": "object",
        "required": [
          "google_client_id",
          "capabilities"
        ],
        "properties": {
          "google_client_id": {
            "type": "string",
            "nullable": true
          },
          "capabilities": {
            "type": "object",
            "required": [
              "google",
              "push",
              "photos",
              "realtime"
            ],
            "properties": {
              "google": {
                "type": "boolean"
              },
              "push": {
                "type": "boolean"
              },
              "photos": {
                "type": "boolean"
              },
              "realtime": {
                "type": "boolean"
              }
            }
          }
        }
      },
      "GoogleProfile": {
        "type": "object",
        "required": [
          "sub",
          "name",
          "email"
        ],
        "properties": {
          "sub": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "picture": {
            "type": "string"
          }
        }
      },
      "Room": {
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "pattern": "^[0-9]{6}$"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "RoomCreate": {
        "type": "object",
        "required": [
          "passcode"
        ],
        "properties": {
          "code": {
            "type": "string",
            "pattern": "^[0-9]{6}$"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1
          },
          "passcode": {
            "type": "string",
            "pattern": "^[0-9]{4,6}$",
            "writeOnly": true
          },
          "nickname": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1
          },
          "google_identity_token": {
            "type": "string"
          }
        }
      },
...
      "AuthSession": {
        "type": "object",
        "required": [
          "room",
          "token",
          "nickname"
        ],
        "properties": {
          "room": {
            "$ref": "#/components/schemas/Room"
          },
          "token": {
            "type": "string"
          },
          "nickname": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1
          },
          "google_profile": {
            "$ref": "#/components/schemas/GoogleProfile"
          }
        }
      },
...
    },
    "securitySchemes": {
      "RoomBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Verified room session. Signature, expiry and room membership are checked."
      },
      "CronBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "CRON_SECRET, never a room session."
      }
    }
  }
}